omiu - move dir
