+++ packages/omiu/src/components/navigation/pagination/_index.css
